% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbourhood.r
\name{balanced_neighbourhood}
\alias{balanced_neighbourhood}
\title{Numerically stable computation of the size of the neighbourhood of a balanced tree}
\usage{
balanced_neighbourhood(n, log = FALSE)
}
\arguments{
\item{n}{an integer giving the number of taxa.}

\item{log}{logical. Whether to return the log of the requested quantity.
Default is \code{FALSE}.}
}
\value{
an integer counting the size of the rSPR graph neighbourhood.
}
\description{
Numerically stable computation of the size of the neighbourhood of a balanced tree
}
