% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clade_utils.r
\name{clade_index}
\alias{clade_index}
\title{Get clade index}
\usage{
clade_index(x, all.splits)
}
\arguments{
\item{x}{the clade as a vector of tip labels}

\item{all.splits}{a list of all clades (splits)}
}
\value{
an index
}
\description{
Get clade index
}
