% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clade_utils.r
\name{clade_indicators}
\alias{clade_indicators}
\title{Construct clade indicators}
\usage{
clade_indicators(ind, L)
}
\arguments{
\item{ind}{indices that should be \code{1}}

\item{L}{size of the indicator vector}
}
\value{
a binary vector with which clades are present (1) and which are absent (0)
}
\description{
Construct clade indicators
}
