% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbourhood.r
\name{gamma_node}
\alias{gamma_node}
\title{Compute the gamma of an internal node}
\usage{
gamma_node(phy, no)
}
\arguments{
\item{phy}{Phylogenetic tree.}

\item{no}{Internal node.}
}
\value{
The corresponding gamma, defined in Song (2003), of the internal node
of the phylogenetic tree.
}
\description{
This function gives the gamma of the internal node in a phylogenetic tree.
}
\examples{

t <- ape::rtree(5, rooted = TRUE)
gamma_node(t, 8)
# Output:  [1] 1

}
\references{
Song, Y. S. (2003). On the combinatorics of rooted binary
phylogenetic trees. \emph{ Annals of Combinatorics, 7(3):365?37}
}
