% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_c.r
\name{get_largest_clade}
\alias{get_largest_clade}
\title{Get the largest clade in a tree}
\usage{
get_largest_clade(phy)
}
\arguments{
\item{phy}{a phylo object.}
}
\value{
a list containing the largest tree and its size
}
\description{
Get the largest clade in a tree
}
