% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_c.r
\name{get_random_clade}
\alias{get_random_clade}
\title{Get a random clade}
\usage{
get_random_clade(phy)
}
\arguments{
\item{phy}{a phylo object}
}
\value{
a list containing a random clade and its size.
}
\description{
Get a random clade
}
