% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transition_matrix.r
\name{make_lazy_MH_matrix}
\alias{make_lazy_MH_matrix}
\title{'Lazify' a Metropolis-Hastings transition matrix}
\usage{
make_lazy_MH_matrix(MHMat, rho)
}
\arguments{
\item{MHMat}{original MH transition matrix}

\item{rho}{stickyness coefficient. The probability of staying put.}
}
\value{
a MH transition matrix of a rho-lazy MH process
}
\description{
'Lazify' a Metropolis-Hastings transition matrix
}
