% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transition_matrix.r
\name{make_MH_matrix}
\alias{make_MH_matrix}
\title{Create the Metropolis-Hastings transition matrix}
\usage{
make_MH_matrix(incidence.mat)
}
\arguments{
\item{incidence.mat}{an incidence matrix (binary entries) of the SPR graph}
}
\value{
a MH transition matrix
}
\description{
Create the Metropolis-Hastings transition matrix
}
