% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbourhood.r
\name{makeLadder}
\alias{makeLadder}
\title{Make a ladder tree}
\usage{
makeLadder(ntaxa)
}
\arguments{
\item{ntaxa}{numerical. An integer giving the number of taxa}
}
\value{
a phylo object containing a ladder treee
}
\description{
Make a ladder tree
}
