% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbourhood.r
\name{max_size}
\alias{max_size}
\title{Maximum rSPR neighbourhood size (i.e. for a balanced tree)}
\usage{
max_size(n)
}
\arguments{
\item{n}{an integer giving the number of taxa.}
}
\value{
an integer counting the number of neighbours in the rSPR graph.
}
\description{
Maximum rSPR neighbourhood size (i.e. for a balanced tree)
}
