% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbourhood.r
\name{min_size}
\alias{min_size}
\title{Minimum rSPR neighbourhood size (i.e. for a ladder tree)}
\usage{
min_size(n)
}
\arguments{
\item{n}{an integer giving the number of taxa.}
}
\value{
an integer counting the number of neighbours in the rSPR graph.
}
\description{
Minimum rSPR neighbourhood size (i.e. for a ladder tree)
}
