% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbourhood.r
\name{neighbour_size_song}
\alias{neighbour_size_song}
\title{Size of the rSPR neighbourhood using the method of Song (2003)}
\usage{
neighbour_size_song(phy)
}
\arguments{
\item{phy}{Phylogenetic tree.}
}
\value{
The number of neighbours of a phylogenetic tree in the rSPR Graph.
}
\description{
Compute the number of neighbours of a phylogenetic tree in the rSPR graph by
the method described in Song (2003).
}
\examples{
t <- ape::rtree(5, rooted = TRUE)
neighbour_size_song(t)
# Output:  [1] 26

}
\references{
Song, Y. S. (2003). On the combinatorics of rooted binary
phylogenetic trees. \emph{ Annals of Com-binatorics, 7(3):365?37}
}
