% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbourhood.r
\name{neighbourhood_size}
\alias{neighbourhood_size}
\title{Compute the neighbourhood size in the rSPR graph}
\usage{
neighbourhood_size(phy)
}
\arguments{
\item{phy}{a phylo object}
}
\value{
an integer counting how many neighbours a tree has in the rSPR graph.
}
\description{
Compute the neighbourhood size in the rSPR graph
}
