% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbourhood.r
\name{neighbourhood_size_matsen}
\alias{neighbourhood_size_matsen}
\title{Compute the neighbourhood size in the rSPR graph using the method from Matsen & Whidden (2017)}
\usage{
neighbourhood_size_matsen(phy)
}
\arguments{
\item{phy}{a phylo object}
}
\value{
an integer counting how many neighbours a tree has in the rSPR graph.
}
\description{
Compute the neighbourhood size in the rSPR graph using the method from Matsen & Whidden (2017)
}
