% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbourhood.r
\name{Nsize_ratio}
\alias{Nsize_ratio}
\title{Numerically stable computation of the ratio of the size of the neighbourhood
of a ladder tree to a balanced tree.}
\usage{
Nsize_ratio(n, log = FALSE)
}
\arguments{
\item{n}{an integer giving the number of taxa.}

\item{log}{logical. Whether to return the log of the requested quantity.
Default is \code{FALSE}.}
}
\value{
an integer counting the size of the rSPR graph neighbourhood.
}
\description{
Numerically stable computation of the ratio of the size of the neighbourhood
of a ladder tree to a balanced tree.
}
