% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rspR.r
\name{rspr}
\alias{rspr}
\title{Compute the rooted SPR (rSPR) distance between two trees}
\usage{
rspr(t1, t2, exact = TRUE)
}
\arguments{
\item{t1}{a phylo object.}

\item{t2}{a phylo object.}

\item{exact}{logical. Whether to compute the exact distance (default is \code{TRUE}) or
an approximation.}
}
\value{
a numerical distance
}
\description{
Compute the rooted SPR (rSPR) distance between two trees
}
\details{
Note that this assumes rspr, \url{https://github.com/cwhidden/rspr}
is installed in the system, ie., at /usr/bin/rspr.
}
