% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rspR.r
\name{rspr_matrix}
\alias{rspr_matrix}
\title{Compute a matrix of rSPR distances}
\usage{
rspr_matrix(trees, type = c("restricted", "onerow", "full"), maxdist = 1)
}
\arguments{
\item{trees}{a multiPhylo object}

\item{type}{if 'restricted', this will compute distances up to \code{maxdist}.
If \code{onerow}, will compute the set of SPR distances from the first tree and
if \code{full} ompute the full matrix (type = "full").
While fun for small data sets, may be impractical for large ones.}

\item{maxdist}{numerical. The maximum distance up to which to compute distances.}
}
\value{
a matrix of distances
}
\description{
Compute a matrix of rSPR distances
}
