% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo_MC.r
\name{run_one_MC}
\alias{run_one_MC}
\title{Run a (phylogenetic) Markov chain from a transition matrix}
\usage{
run_one_MC(Niter, MH_mat, ini.state = NULL)
}
\arguments{
\item{Niter}{Integer number of iterations}

\item{MH_mat}{Transition matrix}

\item{ini.state}{initial state. If \code{NULL}, a state is chosen uniformly.}
}
\value{
A markov chain of tree indices
}
\description{
Run a (phylogenetic) Markov chain from a transition matrix
}
