% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_c.r
\name{safe_f_c}
\alias{safe_f_c}
\title{Safe version of \code{f_c}}
\usage{
safe_f_c(phy, clade)
}
\arguments{
\item{phy}{a phylo object.}

\item{clade}{a list of tip names in a clade.}
}
\value{
a list with two phylo objects: one is the tree without the clade and
the other is the subtree with the clade. Or \code{NA} if
the clade is not compatible with \code{phy}
}
\description{
Safe version of \code{f_c}
}
