% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape_lumping_aux.r
\name{shape_lump}
\alias{shape_lump}
\title{Project (lump) a transition matrix on the space of trees to the space of shapes}
\usage{
shape_lump(transMat, n, ncores = 6)
}
\arguments{
\item{transMat}{Markov transtion matrix on the space of trees}

\item{n}{the number of taxa.}

\item{ncores}{the number of cores to be used.}
}
\value{
a projected (lumped) transition matrix on the space of shapes
}
\description{
Project (lump) a transition matrix on the space of trees to the space of shapes
}
