% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape_lumping_aux.r
\name{vec2symMat}
\alias{vec2symMat}
\title{Create a symmetric matrix from a vector}
\usage{
vec2symMat(x, diag = TRUE, byrow = FALSE)
}
\arguments{
\item{x}{a vector}

\item{diag}{boolean indicating whether to include the diagonal (default is \code{TRUE})}

\item{byrow}{boolean indicating whether to compute the matrix by row.}
}
\value{
a symmetric matrix
}
\description{
Create a symmetric matrix from a vector
}
