% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbourhood.r
\name{compute_gamma}
\alias{compute_gamma}
\title{Sum of gammas of a phylogenetic tree}
\usage{
compute_gamma(phy)
}
\arguments{
\item{phy}{Phylogenetic tree.}
}
\value{
The sum of the gammas from the internal nodes of a phylogenetic tree,
defined in Song (2003).
}
\description{
This function gives the sum of the gammas from a phylogenetic tree.
}
\examples{
t <- ape::rtree(5, rooted = TRUE)
compute_gamma(t)
# Output:  [1] 1

}
\references{
Song, Y. S. (2003). On the combinatorics of rooted binary
phylogenetic trees. \emph{ Annals of Combinatorics, 7(3):365?37}
}
