% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clade_utils.r
\name{get_clades}
\alias{get_clades}
\title{Get the clades in a tree}
\usage{
get_clades(tree, labels)
}
\arguments{
\item{tree}{a phylo object}

\item{labels}{labels for the tips}
}
\value{
a vector with the 2n-1 clades
}
\description{
Get the clades in a tree
}
